% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{addColorMap}
\alias{addColorMap}
\alias{is.CM}
\alias{print.CM}
\alias{plot.CM}
\title{Add Basic Color Map to a Plot}
\usage{
addColorMap(Z, vr = NULL, nc = 11, clr = NULL, z = NULL, border = NA,
  add = T, return = T, ...)

is.CM(x)

\method{print}{CM}(x, digits = 2, ...)

\method{plot}{CM}(x, type = "n", xlab = " ", ylab = " ", xaxt = "n",
  yaxt = "n", bty = "n", ...)
}
\arguments{
\item{Z}{a matrix with the numeric values
to map to a set of colors, or a color
map list object of class 'CM' (the
output for this function).}

\item{vr}{an optional vector giving the lower and
upper boundaries for to use when mapping the
values to colors (defaults to the range for Z).}

\item{nc}{the number of equally-spaced
intervals to use when mapping the
values to colors (default approach needs an odd
number).}

\item{clr}{the color values to which values should
be mapped.}

\item{z}{the endpoints for the intervals to use
when mapping the values to colors (need to be
sequential and equal to the number of colors plus one).}

\item{border}{controls the border for each color square.}

\item{add}{logical; if true, adds the color map to an
existing plot.}

\item{return}{logical; if true, returns a color map list
object.}

\item{...}{Additional inputs for the
\code{\link[graphics]{polygon}} function.}
}
\value{
If indicated, returns a color map list object
  consisting of...
  \itemize{
    \item Z: the original matrix of data.
    \item cmap: a list with z, the endpoints
      of the intervals used to map colors to values,
      and clr, the set of colors to which values
      were assigned.
    \item x: a matrix with the x-axis values for
      each color cell for the \code{polygon} function.
    \item y: a matrix with the y-axis values for
      each color cell for the \code{polygon} function.
    \item clr: a list giving 1) the color for each cell
      in the map, 2) the x-axis position, 3) the y-axis
      position, and 4) the associated value.
    \item xl: the lower and upper intervals for the
      width of the color map.
    \item yl; the lower and upper intervals for
      the height of the color map.
  }
}
\description{
Add Basic Color Map to a Plot
}
\examples{
# Generate a correlation matrix
n = 5 # Number of variables
A = matrix( runif(n^2)*2-1, ncol = n )
# Covariance matrix
S = t(A) \%*\% A
Z = cov2cor( S )

# Create a blank plot
plot( c(0,5), c(0,5), type = 'n', xlab = 'X', ylab = 'Y' )
CM = addColorMap( Z )

# Custom color map
plot( c(0,5), c(0,5), type = 'n', xlab = 'X', ylab = 'Y' )
addColorMap( Z, clr = heat.colors(11), return = F )

# Custom intervals
plot( c(0,5), c(0,5), type = 'n', xlab = 'X', ylab = 'Y' )
z = c( -1, -.75, -.5, 0, .5, .75, 1 ); nc = length(z)
addColorMap( Z, z = z, nc = nc, return = F )

# Create a color map initially, then add to plot
Z = matrix( runif(40), 10, 4 )
CM = addColorMap( Z, clr = heat.colors(11), add = F )
plot( CM )

}
