% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{b}
\alias{b}
\title{Generate Multi-line comments}
\usage{
b(string)
}
\arguments{
\item{string}{A character string with the desired comments.}
}
\description{
A function that allows the user to write multi-line
comments that are not run.
}
\examples{
b("
First line of comments
Second line of comments
")

}

