% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{blankPlot}
\alias{blankPlot}
\title{Generate a Blank Plot}
\usage{
blankPlot(xDim = c(0, 1), yDim = c(0, 1))
}
\arguments{
\item{xDim}{the lower and upper boundaries for the
x-axis.}

\item{yDim}{the lower and upper boundaries for the
y-axis.}
}
\value{
A empty plot.
}
\description{
This function generates a completely blank plot.
}

