% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{colorblindPalette}
\alias{colorblindPalette}
\title{Colorblind Friendly Palette}
\usage{
colorblindPalette()
}
\value{
A vector of 8 hexadecimal strings giving
# colors (starting with grey) that are colorblind
friendly.
}
\description{
Generates a set of 8 colors that are
colorblind friendly.
}
\examples{
clrs = colorblindPalette()
plot( 1:8, 1:8, pch = 15, cex = 4, col = clrs )

}
\references{
See http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette.
}
