% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{commonStats}
\alias{commonStats}
\title{Common Summary Statistics}
\usage{
commonStats(x, type = "M (SD)", digits = 1, pad_left = NULL, na.rm = T)
}
\arguments{
\item{x}{A vector of values.}

\item{type}{The type of summary statistic to compute.
\itemize{
  \item 'M (SD)' for mean and standard deviation.
  \item 'F (\%)' or '\% (F)' for frequency and percentages.
  \item 'Mn to Mx' or 'Mn - Mx' for minimum and maximum.
}}

\item{digits}{The number of digits to round to.}

\item{pad_left}{The minimum number of required characters.
If the observed value is less than what is required,
adds spaces to the left.}

\item{na.rm}{Logical; if \code{TRUE} removes \code{NA} values.}
}
\value{
A character string.
}
\description{
A function to compute nicely formatted common summary statistics.
}
\examples{
# Mean and standard deviation
commonStats( rnorm( 100, 100, 15 ), type = 'M (SD)' )

# Frequency and percentage
commonStats( rbinom( 100, 1, .2 ), type = 'F (\%)' )
commonStats( rbinom( 100, 1, .2 ), type = '\% (F)', digits = 0 )

}
