% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{convertMagnitudeAngle}
\alias{convertMagnitudeAngle}
\title{Conversion to Cartesian Coordinates}
\usage{
convertMagnitudeAngle(H, A, degrees = T)
}
\arguments{
\item{H}{a magnitude.}

\item{A}{an angle.}

\item{degrees}{if true, the angle is assumed to be in
degrees.}
}
\value{
A set of cartesian coordinates.
}
\description{
A function to use the magnitude and angle of a 2D vector
and convert them to cartesian coordinates (assumes the
vector starts at [0,0] ).
}
\examples{
cart = convertMagnitudeAngle( 2, 45 )
print( cart )
}

