% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{createIncrement}
\alias{createIncrement}
\title{Create a Variable with Incremental Unit Intervals}
\usage{
createIncrement(x)
}
\arguments{
\item{x}{a vector of values.}
}
\value{
Returns a vector of matching length to \code{x} with
  incremental unit intervals corresponding to each level of the
  original variable.
}
\description{
A function to take a discrete variable and generate a new variable
with the same number of levels whose values increase in unit
increments. This is useful, for instance, if you have a variable
encoding irregular subject ID numbers that you would like to use
for indexing purposes.
}
\examples{
x = c( 2320, 1038, 3010, 7503 )
print( createIncrement(x) )

}

