% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{customAxes}
\alias{customAxes}
\title{Generate Custom Plot Axes}
\usage{
customAxes(xl, yl, pos = c(1, 2), lnSz = 2, type = 2, label = NULL,
  lbPos = c(2.5, 2.5, 1, 2.5), lbSz = 1.5, inc = c(0, 0, 0, 0),
  axisLabel = NULL, axSz = 1.25, axPos = c(-1, -1, -1, -1),
  prec = 2)
}
\arguments{
\item{xl}{a vector with the lower and upper boundaries for the x-axis.}

\item{yl}{a vector with the lower and upper boundaries for the y-axis.}

\item{pos}{a vector indicating at which sides an axis should be drawn.
The possible sides are:
\enumerate{
  \item Bottom (codes = \code{'Bottom'}, \code{'B'}, \code{'b'},
    \code{'1'}, \code{1});
  \item Left (codes = \code{'Left'}, \code{'L'}, \code{'l'},
    \code{'2'}, \code{2});
  \item Top (codes = \code{'Top'}, \code{'T'}, \code{'t'},
    \code{'3'}, \code{3});
  \item Right (codes = \code{'Right'}, \code{'R'}, \code{'r'},
    \code{'4'}, \code{4});
}
The number of elements in the vector tell the function how many axes
to draw.}

\item{lnSz}{the width of the axis lines.}

\item{type}{the type of axis line to draw. \code{1} or \code{'extend'}
generates axis lines that extend past the boundaries, while \code{2}
or \code{'truncated'} truncates the axis lines at the boundaries.}

\item{label}{the labels for the axes. Must match the length of variable
\code{pos} else a warning is generated and no labels are added.}

\item{lbPos}{the line positions of the labels for the axes.}

\item{lbSz}{the font-size of the labels for the axes.}

\item{inc}{the increments for the axis values. Must match the length of
the variable \code{pos} else a warning is generated and no axis values
are added. Set to \code{0} to suppress the adding of axis values.}

\item{axisLabel}{An optional list matching in length to the variable
\code{pos}, with each element consisting either of a list of the tick
positions and their labels for a given axis, or \code{NULL} to suppress
adding values to that axis.}

\item{axSz}{the size of the font for the axis values.}

\item{axPos}{the line position of the axes.}

\item{prec}{the number of decimal values to report for the axis values.}
}
\description{
A function that, based on the x and y-axis boundaries,
draws lines and labels for a specified set of axes.
}
\examples{
layout( cbind( 1, 2 ) ) # 2 plotting panes
xl = c(0,1); yl = c(0,1); blankPlot(xl,yl) # Empty plot
customAxes( xl, yl, label = c( 'X', 'Y' ), lnSz = 2,
  inc = c(.25,.25), axSz = .75 )
xl = c(.5,2.5); yl = c(-1,1); blankPlot(xl,yl); abline( h = 0 )
customAxes( xl, yl, pos = c(1,4), label = c( 'Conditions', 'Values' ),
  lbSz = 1, lbPos = c(1,1),
  axisLabel = list( list( c(1,2), c('A','B') ), NULL ),
  inc = c(0,.5), axSz = .75 )

}
