% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{defaultAxes}
\alias{defaultAxes}
\title{Add an Axis to a Plot with Default Options}
\usage{
defaultAxes(
  at,
  labels = TRUE,
  side = 1,
  tick = FALSE,
  line = -1.5,
  cex.axis = 1.25,
  ...
)
}
\arguments{
\item{at}{The positions for the tick points.}

\item{labels}{Typically a set of labels of matching length to
\code{at} to be placed at the tick points.}

\item{side}{An integer specifying which side of the plot the
axis is to be drawn on. As per base R, 1 = below, 2 = left,
3 = above, and 4 = right. Defaults to below.}

\item{tick}{Logical; if \code{TRUE} adds tick marks. Default
is \code{FALSE}.}

\item{line}{The number of lines into the margin at which the
axis line will be drawn, if not NA.}

\item{cex.axis}{The size of the text for the labels placed
at the tick points.}

\item{...}{Additional arguments passed on to the \code{\link[graphics]{axis}}
function.}
}
\description{
An extension of the base R function \code{\link[graphics]{axis}}
with default parameters for a subset of the arguments.
}
