% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{degreesRadians}
\alias{degreesRadians}
\title{Conversion between Degrees and Radians}
\usage{
degreesRadians(degrees = NULL, radians = NULL)
}
\arguments{
\item{degrees}{radians equal degrees(pi/180).}

\item{radians}{degrees equal 180*radians/pi.}
}
\value{
The converted values.
}
\description{
A function to convert degrees to radians and vice versa.
}
\examples{
r = degreesRadians( degrees = 45 )
print( r )
d = degreesRadians( radians = .7853982 )
print( d )

}

