% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{densityPoints}
\alias{densityPoints}
\title{Estimate Density for Individual Observations}
\usage{
densityPoints(x, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{additional parameters for the
\code{\link[stats]{density}} function.}
}
\value{
A list with...
  \itemize{
    \item x - the sorted values for the original input.
    \item y - the associated empirical densities.
  }
}
\description{
Given a vector of values, computes the empirical density
for each observation.
}
\examples{
plot(c(-4,4),c(0,.5),type='n',ylab='Density',xlab='z-scores')
x = rnorm( 100 )
dp = densityPoints( x )
points( dp$x, dp$y, pch = 19 )

}
