% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{designCoding}
\alias{designCoding}
\title{Create Design Matrix}
\usage{
designCoding(X, Levels = NULL, Mapping = NULL, type = "Dummy")
}
\arguments{
\item{X}{a vector of categories}

\item{Levels}{the unique values of X. If left unspecified, the
function attemps to extract, but the mappings will be based on
order of appearance.}

\item{Mapping}{For \code{Dummy} and \code{Effects} options,
provides an additional way to indicates which unique values of
X should be mapped to which columns of the design matrix.
For the \code{Coef} option, provides the corresponding weight
values for the unique values of X.}

\item{type}{Current options include...
\itemize{
  \item \code{Dummy} -> Given K levels and N trials, returns a
  N x (K-1) design matrix using dummy coding (i.e. 0 and 1,
  with one variable). Useful for simple effects.
  \item \code{Effects} -> Given K levels and N trials, returns a
  N x (K-1) design matrix using effects coding (i.e. -1 and 1,
  with one variable denoted solely by -1). Useful for comparisons
  against the grand mean.
  \item \code{Intercept} -> Given K levels and N trials, returns a N x K
  design matrix in which each unique level has its own column
  (i.e. K unique intercepts).
  \item \code{Coef} -> Given K levels and N trials, returns a N x 1
  design matrix in which a set of weights (specified via the
  \code{Mapping} variable) are matched to the unique levels of X.
}}
}
\value{
A design matrix.
}
\description{
A function that translates a vector of categories into different
types of design matrices.
}
\examples{
# Default is dummy coding
designCoding( 1:5 )
# Intercept, using mapping option
designCoding( 1:5, Levels = c(3,1,2,4,5), type = 'Intercept' )
# Example of effects coding with the levels reversed
# Note that here, mapping ranges from 0 - 4, not 1 - 5
designCoding( 1:5, Mapping = 4:0, type = 'Effects' )
# Coefficients
set.seed(500)
designCoding( 1:5, Mapping = rnorm(5), type = 'Coef' )

}

