% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{erf}
\alias{erf}
\title{Error Function}
\usage{
erf(x)
}
\arguments{
\item{x}{a vector of values on the real number line.}
}
\value{
a vector of transformed values based on the error function.
}
\description{
Calculates the error function.
}
\examples{
plot( c(-1,1), c(-3,3), type='n', xlab='x', ylab='erf(x)' )
x = seq(-3,3,length=100)
lines(x,erf(x))

}
