% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{EZdiffusion}
\alias{EZdiffusion}
\title{Estimate Parameters for a Two-Boundary Unbiased Wiener Process}
\usage{
EZdiffusion(dat, s = 1)
}
\arguments{
\item{dat}{a vector of inputs:
\enumerate{
  \item Total number of trials;
  \item Number of correct responses;
  \item Mean response time for correct responses;
  \item Variance for corrct response times;
}}

\item{s}{the scaling parameter (typically set to either 0.1 or 1)}
}
\value{
An estimate of the drift rate, boundary separation,
and non-decision time. Bias is fixed to 0.5.
}
\description{
Estimates the drift rate, boundary separation, and
non-decision time for a two-boundary wiener process
with an equidistant starting point using a R function
adapted from Wagenmaker et al. (2007).
}
\examples{
x = c( N = 100, Correct = 80, MRT = .8, VRT = .4^2 )
EZdiffusion( x )
}
\references{
Wagenmakers, E. J., Van Der Maas, H. L., & Grasman, R. P.
  (2007). An EZ-diffusion model for response time and accuracy.
  Psychonomic bulletin & review, 14, 3-22.
  https://doi.org/10.3758/BF03194023.
}
