% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{findMode}
\alias{findMode}
\title{Find the Mode}
\usage{
findMode(x, discrete = F, ...)
}
\arguments{
\item{x}{a vector (can be numeric or a set of strings).}

\item{discrete}{a logical value; if true, an empirical
estimate of the probability mass function is used to
determine the mode. If false, an empirical estimate of
the probability density function is used instead.}

\item{...}{additional options for the \code{density}
function.}
}
\value{
Returns the estimated mode for the vector.
}
\description{
This function determines the mode for a vector.
}
\examples{
x = rnorm(100)
findMode(x)
x = rbinom(10,20,.5)
findMode(x,discrete=T)
x = c( 'cat','dog','cat','mouse')
findMode(x,discrete=T)
}

