% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{findNA}
\alias{findNA}
\title{Function to identify any NA values by row}
\usage{
findNA(x)
}
\arguments{
\item{x}{a matrix or data frame.}
}
\value{
A logical vector with values of
\code{TRUE} for rows with any \code{NA}
values.
}
\description{
This function identifies rows in a matrix
or data frame that contain any NA values.
}
\examples{
x = matrix( rnorm(9), 3, 3 )
x[2,3] = NA
findNA( x )
x = data.frame( A = c( 1, 2, NA ), B = 0 )
findNA( x )

}
