% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{hdi}
\alias{hdi}
\title{Highest Density Interval Estimator}
\usage{
hdi(sampleVec, credMass = 0.95)
}
\arguments{
\item{sampleVec}{a vector of representative values from a probability
distribution.}

\item{credMass}{a scalar between 0 and 1, indicating the mass
within the credible interval that is to be estimated.}
}
\value{
A vector containing the limits of the HDI.
}
\description{
Estimates the highest density interval for a specified coverage
for a sample drawn from a distribution of interest. Useful for
determining credible intervals given a MCMC sample.
}
\section{References}{

Kruschke, J. (2010). Doing Bayesian data analysis: A tutorial
  introduction with R. Academic Press.
}
\examples{
round( qbeta( c(.025,.975), 10, 10 ), 3 ) # True values
round( hdi( rbeta(1000, 10, 10 ) ), 3 ) # Should be close
}

