% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{headerTemplate}
\alias{headerTemplate}
\title{Template for Project Headers}
\usage{
headerTemplate()
}
\value{
Prints the template to the console window, where it
  can be copied and modified by the user.
}
\description{
Prints a template for creating a project header with the
author's name, email contact, and the date.
}
\examples{
headerTemplate()

}
