% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{histogram}
\alias{histogram}
\title{Nicely Formatted Histogram}
\usage{
histogram(
  x,
  col = "grey",
  border = "white",
  main = "",
  xlab = "Values",
  breaks = "FD",
  tck = 0,
  ...
)
}
\arguments{
\item{x}{A vector of values.}

\item{col}{The color used to fill the bars.}

\item{border}{The color of the border of the bars.}

\item{main}{The title of the figure.}

\item{xlab}{The x-axis label.}

\item{breaks}{See \code{\link[graphics]{hist}}.}

\item{tck}{The size of the axis ticks.}

\item{...}{Additional parameters for the
\code{\link[graphics]{hist}} function.}
}
\description{
An extension of the \code{\link[graphics]{hist}}
function that generates a nicely formatted figure with
additional details on common summary statistics
(mean, median, quartiles, and 2.5% to 97.5% end-points).
}
\examples{

histogram( rnorm( 100 ) )

}
