% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{horizLines}
\alias{horizLines}
\title{Add Horizontal Lines to a Plot}
\usage{
horizLines(yval, xl, ...)
}
\arguments{
\item{yval}{a vector of y-axis values at which to
draw the lines.}

\item{xl}{the left and right boundaries for the width of
the lines.}

\item{...}{additional parameters for the
\code{\link[graphics]{segments}} function.}
}
\description{
Draws a set of horizontal lines all of the same
width onto an existing plot.
}
\examples{
plot( rnorm(100), rnorm(100), xlim = c(-3,3) )
horizLines( -2:2, c(-4,4) )

}
