% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{logistic}
\alias{logistic}
\title{Logistic Function}
\usage{
logistic(x)
}
\arguments{
\item{x}{a set of values ( -Inf <= x <= Inf).}
}
\value{
Returns a set of probabilities.
}
\description{
This function applies the logistic function to a set of values
}
\examples{
logistic( c(0,-2.197225,2.1972255,-Inf,Inf) )

}

