% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(p)
}
\arguments{
\item{p}{a set of probabilities (0 <= p <= 1).}
}
\value{
Returns a set of values that now lie between -Inf and Inf.
}
\description{
This function calculates the logit (log of the odds) of a set of probabilities.
}
\examples{
logit( c(.5,.1,.9,0,1) )
}

