% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lowerUpper}
\alias{lowerUpper}
\title{Plotting Range Function}
\usage{
lowerUpper(int, dat)
}
\arguments{
\item{int}{the desired sub-interval over the range.}

\item{dat}{the vector of values over which to determine the boundaries.}
}
\value{
Returns a lower and upper boundary, multiples of the
specified sub-interval.
}
\description{
This function determines the lower and upper boundaries for
a vector of values given a specified sub-interval for
plotting purposes.
}
\examples{
d = density( rnorm(100) )
x.limit = lowerUpper( .5, d$x )
y.limit = lowerUpper( .1, d$y )
plot( x.limit, y.limit, type='n', xlab='Values', ylab='Density' )
lines(d$x,d$y)
}

