% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lsos}
\alias{lsos}
\title{Improved List of Objects}
\usage{
lsos(..., n = 10)
}
\value{
A list of objects in the workspace, along with their
memory usage and dimensions.
}
\description{
A function created by Dirk Eddelbuettel that lists the objects
currently in the workspace along with pertinent information.
}
\examples{
x = rnorm( 100 )
lsos()

}
\references{
Eddelbuettel, D. (2009). Retrieved from http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
