% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mleSE}
\alias{mleSE}
\title{Calculate Standard Errors and Confidence Intervals}
\usage{
mleSE(fit, alpha = 0.95, fnscale = 1)
}
\arguments{
\item{fit}{an optim object in which the option Hessian == T.}

\item{alpha}{the range for the confidence interval (default
is 0.95).}
}
\value{
A list consisting of two objects:
  SE - a vector with the standard errors for a set of parameters.
  CI - a matrix with two rows, the top row contains the lower
       intervals, the bottom row with the upper intervals.
}
\description{
This function calculates the standard errors and confidence
intervals for a set of parameters extracted from an optim
object.
}
\examples{
mle.f = function( par, x ) { -sum( dnorm(x,par[1],par[2],log=T) ) }
x = rnorm(100,5,2)
res = optim( c(0,1), mle.f, x = x, hessian = T )
mle.SE( res )

mle.f = function( par, x ) { sum( dnorm(x,par[1],par[2],log=T) ) }
x = rnorm(100,5,2)
res = optim( c(0,1), mle.f, x = x, hessian = T, control = list( fnscale = -1 ) )
mle.SE( res, alpha = .99, fnscale = -1 )
}

