% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{p_f}
\alias{p_f}
\title{Compute Category Proportions}
\usage{
p_f(x, val)
}
\arguments{
\item{x}{a vector of values.}

\item{val}{a character vector with the set of possible values
that \code{x} can take on.}
}
\value{
Returns a vector of proportions that sum to 1.
}
\description{
Given a set of posssible values, computes the proportion of
values that actually occurred.
}
\examples{
x = rbinom( 100, 1, .7 )
p_f( x, c("0","1") )

# Multiple categories, some categories don't occur
u = runif(100)
x = rep( 'dog', 100 )
x[ u < .33 ] = 'cat'
p_f( x, c("cat","dog","bird" ) )

}

