% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plotTemplate}
\alias{plotTemplate}
\title{Template for Standard Plotting Code}
\usage{
plotTemplate()
}
\value{
Template of an R function to generate a standard plot.
}
\description{
This function prints to the console a
a template for an R function to generate a
standard plot.
}
