% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pow}
\alias{pow}
\title{Raise a Value to a Power}
\usage{
pow(x, a)
}
\arguments{
\item{x}{a numerical value.}

\item{a}{the exponent.}
}
\value{
Returns the result of raising x to
  the power a.
}
\description{
This function raises a value x to a power a.
It is useful for notational purposes,
clearly separting the value and power from
other aspects of an equation.
}
\examples{
x = 2^4
y = pow( 2, 4 )
x == y

# Sometimes it can be hard to determine
# what is being raised to a power
x = 2 * 13 ^ 4 / 8
# This function makes it easier to tell
x = 2 * pow( 13, 4 ) / 8

}
