% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{printTable}
\alias{printTable}
\title{Print a Nicely Formatted Table}
\usage{
printTable(tbl, return = F)
}
\arguments{
\item{tbl}{a formatted data frame.}

\item{return}{logical; if \code{TRUE}, returns
the character vector with each formatted
row.}
}
\value{
If \code{return} is \code{TRUE} returns the
vector of character strings with the formatted output
for each row of the table.
}
\description{
Given a data frame, prints to console
a formated table of the results.
}
\examples{
data( 'mtcars' )
tbl = aggregate( mtcars$mpg, mtcars[,c('cyl','vs')], mean )
tbl$x = round( tbl$x, 1 )
colnames( tbl ) = c( "# of cylinders", "Engine type", "Miles per gallon" )
printTable( tbl )

}
