% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pvalueMC}
\alias{pvalueMC}
\title{Estimate or Format p-values from Monte Carlo Samples}
\usage{
pvalueMC(x, comparison = 0, alternative = "two-sided", digits = 3)
}
\arguments{
\item{x}{either 1) a vector of numeric values (Monte
Carlo samples) or 2) a single p-value.}

\item{comparison}{the comparison point used to compute
the p-value for the Monte Carlo samples.}

\item{alternative}{a character string, either 1) 'two-sided',
2) 'greater', or 3) 'less', indicating the type of
alternativ hypothesis to test. If 'two-sided', uses the
alternativ hypothsis that produces the smallest p-value,
and then adjusts for the two-sided comparison by multiplying
by two.}

\item{digits}{the number of digits to round to when formatting
the numeric p-value}
}
\value{
Either a numeric p-value or a character string,
a nicely formatted version of the p-value.
}
\description{
Given a set of Monte Carlo samples, estimates a p-value
from the proportion of values that fall above or below
a comparison point. If \code{string} is \code{TRUE},
takes a numeric p-value and converts it into a
formatted character string, either 'p = ...' or
'p < ...'.
}
\examples{
x = rnorm( 1000 )
p = pvalueMC( x )
print( pvalueMC( p ) )
p = pvalueMC( x, alternative = 'greater' )
print( pvalueMC( p ) )
p = pvalueMC( x, alternative = 'less' )
print( pvalueMC( p, digits = 4 ) )

}
