% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quick_doc_template}
\alias{quick_doc_template}
\title{Template for Documentation}
\usage{
quick_doc_template()
}
\value{
Prints the template to the console window, where it
  can be copied and modified by the user.
}
\description{
Prints a template for an informal way to document R functions
to the console.
}
\examples{
quick_doc_template()

}

