% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quickDocTemplate}
\alias{quickDocTemplate}
\title{Template for Documentation}
\usage{
quickDocTemplate()
}
\value{
Prints the template to the console window, where it
  can be copied and modified by the user.
}
\description{
Prints a template for an informal way to document R functions
to the console.
}
\examples{
quickDocTemplate()

}
