% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sem}
\alias{sem}
\title{Standard Error of the Mean}
\usage{
sem(x)
}
\arguments{
\item{x}{a vector of values.}
}
\value{
Returns the estimated standard error of the mean
  based on the values of x
}
\description{
This function calculates the standard error of the mean.
}
\examples{
sem( rnorm(1000) )
}

