% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{simpleBootstrap}
\alias{simpleBootstrap}
\title{Simple Bootstrap Function}
\usage{
simpleBootstrap(x, T_x = mean, nRep = 1000, ...)
}
\arguments{
\item{x}{a numeric vector, or a long-form matrix or data frame.}

\item{T_x}{a function to compute a desired test statistic
(default is the mean).}

\item{nRep}{the number or replicates to sample.}

\item{...}{additional parameters for the test statistic function.}
}
\value{
Returns a list containing the test statistics for each
  replicate, along with the test statistic for the original
  set of data.
}
\description{
A function for computing a test statistic over
multiple replicates of a set of data. Replicates
are created by sampling from the original data with
replacement.
}
\examples{
# Vector
x = rnorm( 100 )
bootstrap = simpleBootstrap( x, T_x = median )
quantile( bootstrap$replicates, c( .16, .84 ) )

# Matrix
x = cbind( rlnorm( 100, -2, .5 ), rbeta( 100, 4, 1 ) )
bootstrap = simpleBootstrap( x, T_x = colMeans )
apply( bootstrap$replicates, 1, quantile, prob = c( .16, .84 ) )

}
