% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{simplePiecewiseReg}
\alias{simplePiecewiseReg}
\title{Fit a Basic Piecewise Regresssion}
\usage{
simplePiecewiseReg(x, y, breaks, ...)
}
\arguments{
\item{x}{the independent (predictor) variable.}

\item{y}{the dependent variable.}

\item{breaks}{a vector of break points (in the same
scale as the independent variable).}

\item{...}{additional parameters for the
\code{\link[stats]{lm}} function.}
}
\value{
A list with the output from the
  \code{\link[stats]{lm}} function,
  a data frame with the dummy coding used to
  construct the  piecewise regression, and
  x and y-axis values that can be used for
  plotting the estimated line segments.
}
\description{
Given an indepdent and dependent variable, along
with a vector of break points, fits a basic
piecewise regression to the data.
}
\examples{
# Define generating parameters
beta = c( 1, 1, 1, -1 )
sigma = .5

# Define break point
breaks = 0

# Simulate data
x = runif( 100, -1, 1 ) # Predictor
# Define design matrix
X = matrix( 1, 100, 2 )
X[ x > breaks, 1] = 0; X[ x <= breaks, 2] = 0
X = cbind( X, X*x )
colnames( X ) = c( 'I1', 'I2', 'S1', 'S2' )
# Generate dependent variable
y = as.vector( X \%*\% cbind( beta ) )
y = y + rnorm( 100, 0, sigma )

# Fit piecewise regression
fit = simplePiecewiseReg( x, y, breaks )
summary( fit$lm )

Plot results
plot( x, y, pch = 19 )
lines( fit$xa, fit$ya )

}
