% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{standardize}
\alias{standardize}
\title{Custom Function to Standardize a Variable}
\usage{
standardize(x, reverse = F)
}
\arguments{
\item{x}{a numerical value.}

\item{reverse}{logical; if true, the function
extracts the 'mean' and 'scale' attributes
and reverses the standardization.}
}
\value{
If reverse is FALSE, returns the standardized
values for x, with attributes giving the mean and scale;
otherwise, returns the raw, unstandardized values
of x.
}
\description{
This function centers and scales a variable,
and adds attributes that allow the function
to convert the standardized variable back
into its original scale.
}
\examples{
x = runif( 30 )
# Standardize the variable
z = my_standardize( x )
# Unstandardize the variable
y = my_standardize( x, reverse = T )
all( x == y )

}
