% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tick}
\alias{tick}
\alias{tock}
\title{Functions to Evaluate Run Times}
\usage{
tick()

tock()
}
\value{
The function \code{tick} creates a global
variable 'run_time' with the current system time.
The function \code{tock} then updates 'run_time'
with the difference between the current system
time and the previous value of 'run_time'.
}
\description{
The functions \code{tick} and \code{tock} can be
used to roughly estimate the run time of a segment
of code.
}
\examples{
tick()
Sys.sleep(2.5)
tock()
print( run_time )

}
