% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{vertLines}
\alias{vertLines}
\title{Add Vertical Lines to a Plot}
\usage{
vertLines(xval, yl, ...)
}
\arguments{
\item{xval}{a vector of x-axis values at which to
draw the lines.}

\item{yl}{the bottom and top boundaries for the height of
the lines.}

\item{...}{additional parameters for the
\code{\link[graphics]{segments}} function.}
}
\description{
Draws a set of vertical lines all of the same
height onto an existing plot.
}
\examples{
plot( rnorm(100), rnorm(100), ylim = c(-3,3) )
vertLines( -2:2, c(-4,4) )

}
