% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{violinPlot}
\alias{violinPlot}
\title{Draw a Violin Plot}
\usage{
violinPlot(x, pos, scaleH = 0.5, interval = list(), ...)
}
\arguments{
\item{x}{the vector of values used to create the
violin plot.}

\item{pos}{the x-axis position at which to draw the
violin plot}

\item{scaleH}{the maximum half-width of the violin plot.}

\item{interval}{allows for extra options governing ways
to restrict the violin plot to a  desired sub-interval.
\describe{
  \item{\code{crit}}{a critical value or pair of values
    that determines the sub-interval to plot over.}
  \item{\code{type}}{when \code{crit} takes on a single
    value, 'greater' restricts the sub-interval to values
    larger than \code{crit},'less' to values below
    \code{crit}.}
  \item{\code{out}}{a logical value; if true, the
    function returns the proportion of values above,
    below, or within \code{crit}.}
}}

\item{...}{additional plotting parameters for \code{polygon}.}
}
\value{
If \code{out} is true, returns the proportion of values
  in \code{x} above, below, or within the interval defined by
  \code{crit}.
}
\description{
Adds a violin plot to an already existing figure.
}
\examples{
x = rnorm(100)
blankPlot(yDim=c(-6,6))
crit = mean(x)
violinPlot( x, .5, interval = list( crit = crit ), col = 'grey',
  border = NA )
violinPlot( x, .5, lwd = 2 )
}

