% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{correlationHeatmap}
\alias{correlationHeatmap}
\title{Plot a Heatmap for a Correlation Matrix}
\usage{
correlationHeatmap(
  df,
  ttl = "Correlation matrix",
  labels = NULL,
  new = T,
  lyt = NULL,
  gradient = c("orange", "blue"),
  txtSz = 1.25,
  mc_adjust = "BH",
  cut_off = 0.05,
  H = 20/3,
  W = 25/3
)
}
\arguments{
\item{df}{a data frame with all variables to include
in the correlation matrix.}

\item{ttl}{an optional title for the figure.}

\item{labels}{the labels for the rows/columns.}

\item{new}{logical; if \code{TRUE} generates a
new plotting window.}

\item{lyt}{an optional matrix specifying the
layout of the main panel (1) versus
the side panel (2) with the color gradient.}

\item{gradient}{the final end colors for
the negative and positive correlations, respectively.}

\item{txtSz}{the size of the text in the figure.}

\item{mc_adjust}{the method to use when correcting for
multiple comparisons (see ).}

\item{cut_off}{cut-off for significance.}

\item{H}{the height to use if a new plotting window is
generated.}

\item{W}{the width to use if a new plotting window is
generated.}
}
\value{
A heatmap for the upper-triangle portion of the
  correlation matrix.
}
\description{
Generates a heatmap of the upper triangle of a
correlation matrix.
}
\examples{
# Load data
data("mtcars")
my_data <- mtcars[, c(1,3,4,5,6,7)]
# Generate heatmap
correlationHeatmap( my_data, labels = colnames(my_data) )

}
