% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{errorBars}
\alias{errorBars}
\title{Draw Error Bars}
\usage{
errorBars(pos, limits, flip = F, ...)
}
\arguments{
\item{pos}{a single value or a vector of N values, indicating
the position(s) at which an error bar should be drawn.}

\item{limits}{a vector of 2 values or a 2 x N matrix giving the
lower and upper limits, respectively, of the uncertainty
intervals.}

\item{flip}{a logical value. If true, bars are drawn horizontally
instead of vertically. In this case \code{pos} denotes the
position(s) on the y-axis. Otherwise, \code{pos} denotes the
position(s) on the x-axis.}

\item{...}{additional plotting parameters for the
\code{\link[graphics]{arrows}} function.}
}
\description{
Adds error bars to an already existing plot.
}
\examples{
# Example of 95\% versus 68\% intervals for standard normal
plot( c(0,1),c(-6,6),type='n',xaxt='n',xlab=' ',ylab='z-scores' )
pos = c(.25,.75)
limits = cbind( qnorm( c(.025,.975) ), qnorm( c(.16,.84) ) )
errorBars( pos, limits, length = .05, lwd = 2 )
text( pos, limits[2,], c("95\%","68\%"), pos = 3 )
abline(h=0)

}
