% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{higherOrderMoments}
\alias{higherOrderMoments}
\title{Calculate the First Four Moments}
\usage{
higherOrderMoments(x)
}
\arguments{
\item{x}{a numerical vector of values.}
}
\value{
A list consisting of...
  \itemize{
    \item \code{n}: The number of observations.
    \item \code{mean}: The mean of the data (The 1st moment).
    \item \code{var}: The variance of the data (The 2nd moment).
    \item \code{sd}: The standard deviation of the data.
    \item \code{skew}: The skewness of the data (The 3rd moment).
    \item \code{kurtosis}: The kurtosis of the data (The
      fourth moment).
  }
}
\description{
A function to compute the first four moments (i.e., the mean,
variance, skew, and kurtosis) for a vector of data.
}
\section{References}{

Terriberry, T. B. (2007). Computing Higher-Order Moments Online.
}
\examples{
x = rnorm( 100 )
higherOrderMoments( x )

}

